feat(web): add demo route and fullscreen chat layout
Split the legacy Demo into a hash route (/#demo) and keep the default app focused on a fullscreen ChatContainer. Also fix the Live2D right toolbar panel to open to the left.
